requirements.txt
gradio>=3.42.0
requests>=2.31.0
Pillow>=9.5.0
numpy>=1.23.0
tqdm>=4.66.0
opencv-python>=4.9.0.80
facexlib>=0.3.0
nudenet>=2.0.2
# PyTorch (torch, torchvision, torchaudio) se instalará por separado con soporte CUDA mediante setup_env.ps1
onnxruntime>=1.16.3 